Fixed - Logging issue on startup today. Configuration couldn't find the correct files. Moved - To fix and clean-up, I moved the src and tdameritrade folders out of the stock_watch_app folder and removed that entirely.
