+++ src/handlers/api_handler.py
@@ -1,6 +1,6 @@
 import requests
-import stock_watch_app.src.extends.logger as logging
-from stock_watch_app.src.models.api_config import APIConfig
+import src.extends.logger as logging
+from src.models.api_config import APIConfig
 
 logger = logging.get_logger(__name__)
