+++ tdameritrade/models/types/stock_index_type.py
