+++ tdameritrade/models/types/value_change_type.py
